Fix ptype ineligibility check
Fixes recently added check for ptype. Previously, although we were
checking ptype to determine eligibility, we were only doing so *after*
the test hold failed. Because the barred ptypes are technically able to
place holds (which is the reason for this work) the test hold will not
fail for patrons with barred ptypes (unless there are other issues with
their card. Thus, this patch rearranges the order of checks to ensure
that both the test hold and ptype check occur at the outset.

https://jira.nypl.org/browse/SCC-1638