Reduce false alarms due to logging normal errors
The app attempts to place a hold on a non existant item to determine if
the card has issues. If the error received is anything other than the
most common expected error, the app logs the error out. Because the
error contains the phrase "Error", this is triggering our error alarms
any time the eligibility service encounters a patron with card issues.
This fix moves that logging into the debug log (and includes context).